serializer package restructure
