+++ trustpoint/pki/download/certificate.py
@@ -9,12 +9,12 @@ from django.http import HttpResponse, Http404
 from docutils.nodes import field
 
 from pki.models import CertificateModel
-from pki.serialization.serializer import CertificateSerializer, CertificateCollectionSerializer
+from pki.serializer import CertificateSerializer, CertificateCollectionSerializer
 
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from pki.serialization.serializer import CertificateSerializer, CertificateCollectionSerializer
+    from pki.serializer import CertificateSerializer, CertificateCollectionSerializer
 
 
 class CertificateFileFormat(enum.Enum):
